Split up sensors module and fixed arg order
