+++ enginecore/enginecore/state/assets.py
@@ -25,7 +25,7 @@ from circuits import Component, handler
 import enginecore.state.state_managers as sm
 from enginecore.state.asset_definition import register_asset, SUPPORTED_ASSETS
 from enginecore.state.agents import IPMIAgent, SNMPAgent, StorCLIEmulator
-from enginecore.state.sensors import SensorRepository
+from enginecore.state.sensor.repository import SensorRepository
 
 PowerEventResult = namedtuple("PowerEventResult", "old_state new_state asset_key asset_type")
 PowerEventResult.__new__.__defaults__ = (None,) * len(PowerEventResult._fields)
