+++ enginecore/enginecore/state/sensor/sensor.py
@@ -1,4 +1,6 @@
-"""Aggregates sensor management tools """
+"""Sensor provides access to ipmi/bmc sensors & manages thermal relationships between
+sensors & sensors, sensors & storage components, cpu & sensors
+"""
 
 import os
 import threading
@@ -9,24 +11,6 @@ import operator
 
 import enginecore.state.state_managers as sm 
 from enginecore.model.graph_reference import GraphReference
-
-
-class SensorFileLocks():
-    """File locks for sensor files for safe access"""
-
-    def __init__(self):
-        self._s_file_locks = {}
-
-    def __str__(self):
-        return str(self._s_file_locks)
-
-    def add_sensor_file_lock(self, sensor_name):
-        """Add new lock"""
-        self._s_file_locks[sensor_name] = threading.Lock()
-
-    def get_lock(self, sensor_name):
-        """Get file lock by sensor name"""
-        return self._s_file_locks[sensor_name]
 
 
 class Sensor():
@@ -48,7 +32,7 @@ class Sensor():
         self._th_cpu_t = None
 
         self._th_sensor_t_name_fmt = "({event})s:[{source}]->t:[{target}]"
-        self._th_storage_t_name_fmt = "({event})s:[{source}]->STORAGE:t:[{ctrl}/{target}]"
+        self._th_storage_t_name_fmt = "({event})s:[{source}]->STORAGE:t:[c{ctrl}/{target}]"
 
         self._th_cpu_t_name_fmt = "s:[cpu_load]->t:[{target}]"
 
@@ -65,7 +49,7 @@ class Sensor():
         self._s_file_locks = s_locks
         self._s_thermal_event = threading.Event()
 
-  
+
 
     def __str__(self):
         with self._graph_ref.get_session() as session:
@@ -232,8 +216,9 @@ class Sensor():
         with self._graph_ref.get_session() as session:
             while True:
                 logging.info('\n\nstorage: s:{}, t:{}, e:{}\n\n'.format(
-                    controller, target, event    
+                    controller, target, event
                 ))
+                GraphReference.add_to_cv_temperature(session, self._server_key, controller, target, 0)
                 time.sleep(5)
 
 
@@ -329,7 +314,7 @@ class Sensor():
         if cv in self._th_cv_t:
             raise ValueError('Thread already exists')
 
-        self._launch_thermal_cv_thread(cv, controller, event)
+        self._launch_thermal_cv_thread(controller, cv, event)
 
 
     def add_sensor_thermal_impact(self, target, event):
@@ -419,114 +404,3 @@ class Sensor():
             
             off_value = self._s_specs['offValue'] if 'offValue' in self._s_specs else 0
             filein.write(str(default_value if 'defaultValue' in self._s_specs else off_value))
-
-   
-
-class SensorRepository():
-    """A sensor repository for a particular IPMI device"""
-
-    def __init__(self, server_key, enable_thermal=False):
-        self._server_key = server_key
-        
-        self._graph_ref = GraphReference()
-        self._sensor_file_locks = SensorFileLocks()
-
-        self._sensor_dir = os.path.join(
-            sm.StateManager.get_temp_workplace_dir(),
-            str(server_key),
-            'sensor_dir'
-        )
-
-        self._sensors = {}
-
-        with self._graph_ref.get_session() as session:
-            sensors = GraphReference.get_asset_sensors(session, server_key)
-            for sensor_info in sensors:
-                sensor = Sensor(self._sensor_dir, server_key, sensor_info, self._sensor_file_locks)
-                self._sensors[sensor.name] = sensor
-
-        if enable_thermal:
-            self._load_thermal = True
- 
-            if not os.path.isdir(self._sensor_dir):
-                os.mkdir(self._sensor_dir)
-
-            for s_name in self._sensors:
-                self._sensors[s_name].set_to_defaults()
-
-
-    def __str__(self):
-        
-        repo_str = []
-        repo_str.append("Sensor Repository for Server {}".format(self._server_key))
-        repo_str.append(" - files for sensor readings are located at '{}'".format(self._sensor_dir))
-
-        return '\n\n'.join(repo_str + list(map(lambda sn: str(self._sensors[sn]), self._sensors)))
-
-
-    def enable_thermal_impact(self):
-        """Set thermal event switch """
-        list(map(lambda sn: self._sensors[sn].enable_thermal_impact(), self._sensors))
-
-
-    def disable_thermal_impact(self):
-        """Clear thermal event switch"""
-        list(map(lambda sn: self._sensors[sn].disable_thermal_impact(), self._sensors))
-
-
-    def shut_down_sensors(self):
-        """Set all sensors to offline"""
-        for s_name in self._sensors:
-            sensor = self._sensors[s_name]
-            if sensor.group != 'temperature':
-                sensor.set_to_off()
-
-        self.disable_thermal_impact()
-        
-
-    def power_up_sensors(self):
-        """Set all sensors to online"""
-        for s_name in self._sensors:
-            sensor = self._sensors[s_name]
-            if sensor.group != 'temperature':
-                sensor.set_to_defaults()
-        self.enable_thermal_impact()
-
-
-    def get_sensor_by_name(self, name):
-        """Get a specific sensor by name"""
-        return self._sensors[name]
-
-
-    def adjust_thermal_sensors(self, old_ambient, new_ambient):
-        """Indicate an ambient update"""
-        
-        for s_name in self._sensors:
-            sensor = self._sensors[s_name]
-            if sensor.group == 'temperature':
-                with self._sensor_file_locks.get_lock(sensor.name):
-                    
-                    old_sensor_value = int(sensor.sensor_value)
-                    new_sensor_value = old_sensor_value - old_ambient + new_ambient if old_sensor_value else new_ambient
-
-                    logging.info(
-                        "Sensor:[%s] - value will be updated from %s° to %s° due to ambient changes (%s° -> %s°)",
-                        sensor.name,
-                        old_sensor_value,
-                        new_sensor_value,
-                        old_ambient,
-                        new_ambient
-                    )
-
-                    sensor.sensor_value = int(new_sensor_value)
-
-        if self._load_thermal is True:
-            self._load_thermal = False
-
-            for s_name in self._sensors:
-                self._sensors[s_name].start_thermal_impact()
-
-    @property
-    def sensor_dir(self):
-        """Get temp IPMI state dir"""
-        return self._sensor_dir
